+++ engine/src/main/kotlin/engine/conditions/SignConditions.kt
@@ -13,6 +13,8 @@ fun Expression.isDefinitelyNotNegative(): Boolean = isDefinitely(Sign.NON_NEGATI
 
 fun Expression.isDefinitelyNegative() = isDefinitely(Sign.NEGATIVE)
 
+fun Expression.isDefinitelyPositive() = isDefinitely(Sign.POSITIVE)
+
 fun Expression.isDefinitely(wantedSign: Sign): Boolean {
     val sign = signOf()
     return when {