+++ methods/src/main/kotlin/methods/integerroots/IntegerRootsExplanation.kt
@@ -119,6 +119,48 @@ enum class IntegerRootsExplanation : CategorisedMetadataKey {
      */
     SimplifySquareRootWithASquareFactorRadicand,
 
+    /**
+     * Simplify an expression of the form sqrt[a +/- b sqrt[n]]
+     * when a +/- b sqrt[n] can be written as a square.
+     *
+     * E.g. sqrt[11 - 6sqrt[2]] --> sqrt[(3 - sqrt[2]) ^ 2]
+     *                          --> 3 - sqrt[2]
+     */
+    SimplifySquareRootOfIntegerPlusSurd,
+
+    /**
+     * Write an expression of the form a +/- b sqrt[n] as (x +/- y sqrt[n]) ^ 2 by writing
+     * a system of equations for x and y and finding a solution to it.
+     *
+     * E.g. 11 - 6 sqrt[2] --> (3 - sqrt[2]) ^ 2
+     */
+    WriteIntegerPlusSquareRootAsSquare,
+
+    /**
+     * In the context of factoring a + b sqrt[n] as a square, write an equation in x and y,
+     * deduce a system of two equations that x and y my satisfy and find integer solutions
+     * for x and y.
+     *
+     * E.g. in the context of factoring 11 - 6sqrt[2]
+     *     (x + y sqrt[2])^2 = 11 - 6sqrt[2]
+     * --> x^2 + 2xy sqrt[2] + 2y^2 = 11 - 6 sqrt[2]
+     * --> x^2 + 2y^2 = 11 AND 2xy sqrt[2] = -6 sqrt[2]
+     * --> x^2 + 2y^2 = 11 AND xy = -3
+     * --> x = 3 AND y = -1
+     */
+    WriteEquationInXAndYAndSolveItForFactoringIntegerPlusSurd,
+
+    /**
+     * In the context of factoring  a + b sqrt[n] = (x + y sqrt[2])^2, given that integer solutions
+     * have been found for x and y, substituted them back in order to find the square form for
+     * a + b sqrt[n].
+     *
+     * E.g. in the context of factoring 11 - 6sqrt[2] in the form (x + y sqrt[2])^2,
+     * we found x= 3 AND y = -1.  So the form for 11 - 6sqrt[2] as a square will be:
+     *
+     *     3 - sqrt[2]
+     */
+    SubstituteXAndYorFactoringIntegerPlusSurd,
     ;
 
     override val category = "IntegerRoots"