+++ methods/src/main/kotlin/methods/integerroots/IntegerRootsPlans.kt
@@ -1,28 +1,44 @@
 package methods.integerroots
 
+import engine.expressions.Constants
+import engine.expressions.Extractor
 import engine.expressions.Product
 import engine.expressions.Root
 import engine.expressions.SquareRoot
+import engine.expressions.Variable
 import engine.expressions.containsRoots
+import engine.expressions.equationOf
+import engine.expressions.powerOf
+import engine.expressions.productOf
+import engine.expressions.simplifiedProductOf
+import engine.expressions.squareRootOf
+import engine.expressions.sumOf
 import engine.methods.CompositeMethod
 import engine.methods.RunnerMethod
 import engine.methods.plan
 import engine.methods.stepsproducers.applyToFactors
 import engine.methods.stepsproducers.steps
+import engine.methods.taskSet
 import engine.patterns.AnyPattern
 import engine.patterns.SignedIntegerPattern
 import engine.patterns.UnsignedIntegerPattern
+import engine.patterns.commutativeSumOf
 import engine.patterns.condition
 import engine.patterns.integerOrderRootOf
 import engine.patterns.powerOf
 import engine.patterns.rootOf
 import engine.patterns.squareRootOf
+import engine.patterns.withOptionalIntegerCoefficient
 import engine.steps.metadata.metadata
 import engine.utility.isPowerOfDegree
+import methods.equations.EquationsRules
+import methods.equationsystems.EquationSystemsRules
 import methods.general.GeneralRules
 import methods.integerarithmetic.IntegerArithmeticPlans
 import methods.integerarithmetic.IntegerArithmeticRules
 import methods.integerarithmetic.simplifyIntegersInExpression
+import methods.polynomials.PolynomialsPlans
+import methods.solvable.SolvableRules
 
 enum class IntegerRootsPlans(override val runner: CompositeMethod) : RunnerMethod {
 
@@ -210,6 +226,8 @@ enum class IntegerRootsPlans(override val runner: CompositeMethod) : RunnerMetho
     ),
 
     SimplifySquareRootWithASquareFactorRadicand(simplifySquareRootWithASquareFactorRadicand),
+
+    SimplifySquareRootOfIntegerPlusSurd(simplifySquareRootOfIntegerPlusSurd),
 }
 
 val cancelRootOfPower = steps {
@@ -261,4 +279,83 @@ val simplifySquareRootWithASquareFactorRadicand = plan {
         apply(IntegerRootsRules.MoveSquareFactorOutOfRoot)
         applyTo(IntegerRootsPlans.SimplifyIntegerRoot) { it.firstChild }
     }
+}
+
+val simplifySquareRootOfIntegerPlusSurd = plan {
+    explanation = Explanation.SimplifySquareRootOfIntegerPlusSurd
+    pattern = squareRootOf(
+        commutativeSumOf(
+            UnsignedIntegerPattern(),
+            withOptionalIntegerCoefficient(squareRootOf(UnsignedIntegerPattern())),
+        ),
+    )
+
+    steps {
+        applyTo(writeIntegerPlusSurdAsSquare) { it.firstChild }
+        apply(GeneralRules.CancelRootIndexAndExponent)
+    }
+}
+
+val writeIntegerPlusSurdAsSquare = taskSet {
+    explanation = Explanation.WriteIntegerPlusSquareRootAsSquare
+    val integer = UnsignedIntegerPattern()
+    val radicand = UnsignedIntegerPattern()
+    val root = withOptionalIntegerCoefficient(squareRootOf(radicand))
+
+    // a + b sqrt[c]
+    pattern = commutativeSumOf(integer, root)
+
+    tasks {
+        // E.g. 11 - 6sqrt[2]
+
+        // Task 1 - solve for x and y
+        // (x + y sqrt[2])^2 = 11 - 6sqrt[2]
+        // --> x^2 + 2y^2 + 2xy sqrt[2] = 11 - 6 sqrt[2]
+        // --> x^2 + 2y^2 = 11 AND 2xy sqrt[2] = -6 sqrt[2]
+        // --> x^2 + 2y^2 = 11 AND xy = -3
+        // Now guess two numbers that multiply to -3
+        // --> x = 3 AND y = -1
+
+        val xVar = Variable("x")
+        val yVar = Variable("y")
+
+        val squareInXAndY = powerOf(sumOf(xVar, productOf(yVar, squareRootOf(get(radicand)))), Constants.Two)
+        val findXAndY = task(
+            startExpr = equationOf(squareInXAndY, expression),
+            explanation = metadata(
+                Explanation.WriteEquationInXAndYAndSolveItForFactoringIntegerPlusSurd,
+                squareInXAndY,
+            ),
+            context = context.copy(solutionVariables = listOf("x", "y")),
+            stepsProducer = findXAndYSteps,
+        ) ?: return@tasks null
+
+        // Task 2 - substitute x= 3 and y = 1 into (x + y sqrt[2]) ^ 2
+        // (3 - sqrt[2]) ^ 2
+
+        val solution = findXAndY.result
+        val x = solution.firstChild.secondChild
+        val y = solution.secondChild.secondChild
+
+        val square = powerOf(sumOf(x, simplifiedProductOf(y, squareRootOf(get(radicand)))), Constants.Two)
+
+        task(
+            startExpr = square,
+            explanation = metadata(Explanation.SubstituteXAndYorFactoringIntegerPlusSurd),
+        )
+        allTasks()
+    }
+}
+
+private val findXAndYSteps = steps {
+    applyTo(PolynomialsPlans.ExpandPolynomialExpression) { it.firstChild }
+    apply(EquationsRules.SplitEquationWithRationalVariables)
+    applyTo(Extractor { it.secondChild }) {
+        whilePossible(SolvableRules.CancelCommonFactorOnBothSides)
+        optionally {
+            apply(SolvableRules.FindCommonIntegerFactorOnBothSides)
+            apply(SolvableRules.CancelCommonFactorOnBothSides)
+        }
+    }
+    apply(EquationSystemsRules.GuessIntegerSolutionsOfSystemContainingXYEqualsInteger)
 }