+++ methods/src/test/kotlin/methods/equations/SolveRationalEquationTest.kt
@@ -900,7 +900,7 @@ class SolveRationalEquationTest {
             fromExpr = "[12 / [x ^ 2] - 9] = sqrt[2] - [2 / x + 3]"
             // on one of the good days, we will be able to simplify the below
             // result to "3 + 2sqrt[2]"
-            toExpr = "SetSolution[x : {[sqrt[2] + sqrt[38 + 12 sqrt[2]] / 2]}]"
+            toExpr = "SetSolution[x : {sqrt[2] + 3}]"
             explanation {
                 key = EquationsExplanation.SolveEquation
             }
@@ -917,7 +917,7 @@ class SolveRationalEquationTest {
 
                 step {
                     fromExpr = "[12 / [x ^ 2] - 9] = sqrt[2] - [2 / x + 3]"
-                    toExpr = "SetSolution[x: {[sqrt[2] - sqrt[38 + 12 sqrt[2]] / 2], [sqrt[2] + sqrt[38 + 12 sqrt[2]] / 2]}]"
+                    toExpr = "SetSolution[x: {-3, sqrt[2] + 3}]"
                     explanation {
                         key = EquationsExplanation.SolveRationalEquation
                     }
@@ -953,8 +953,8 @@ class SolveRationalEquationTest {
                     }
 
                     step {
-                        fromExpr = "x = [2 sqrt[2] + (-2 sqrt[19 + 6 sqrt[2]]) sqrt[2] / 4] OR x = [2 sqrt[2] + (2 sqrt[19 + 6 sqrt[2]]) sqrt[2] / 4]"
-                        toExpr = "SetSolution[x: {[sqrt[2] - sqrt[38 + 12 sqrt[2]] / 2], [sqrt[2] + sqrt[38 + 12 sqrt[2]] / 2]}]"
+                        fromExpr = "x = [2 sqrt[2] + (-(2 + 6 sqrt[2])) sqrt[2] / 4] OR x = [2 sqrt[2] + ((2 + 6 sqrt[2])) sqrt[2] / 4]"
+                        toExpr = "SetSolution[x : {-3, sqrt[2] + 3}]"
                         explanation {
                             key = EquationsExplanation.SolveEquationUnion
                         }
@@ -964,7 +964,7 @@ class SolveRationalEquationTest {
 
             task {
                 taskId = "#3"
-                startExpr = "SetSolution[x : {[sqrt[2] + sqrt[38 + 12 sqrt[2]] / 2]}]"
+                startExpr = "SetSolution[x : {sqrt[2] + 3}]"
                 explanation {
                     key = EquationsExplanation.SomeSolutionsDoNotSatisfyConstraint
                 }